Explicitly instantiate some templates (#2366)
