+++ src/test/lib/operators/join_hash_test.cpp
@@ -110,12 +110,13 @@ TEST_F(OperatorsJoinHashTest, DeepCopy) {
 
 TEST_F(OperatorsJoinHashTest, RadixBitCalculation) {
   // Simple tests to check that side switching and zero-sizes work.
-  EXPECT_EQ(JoinHash::calculate_radix_bits<int32_t>(1, 0, JoinMode::Inner), 0ul);
-  EXPECT_EQ(JoinHash::calculate_radix_bits<int32_t>(0, 1, JoinMode::Inner), 0ul);
-  EXPECT_EQ(JoinHash::calculate_radix_bits<int32_t>(0, 0, JoinMode::Inner), 0ul);
-  EXPECT_EQ(JoinHash::calculate_radix_bits<int32_t>(1, 1, JoinMode::Inner), 0ul);
-  EXPECT_TRUE(JoinHash::calculate_radix_bits<int32_t>(std::numeric_limits<size_t>::max(),
-                                                      std::numeric_limits<size_t>::max(), JoinMode::Inner) > 0ul);
+  EXPECT_EQ(JoinHash::calculate_radix_bits(1, 0, JoinMode::Inner), 0ul);
+  EXPECT_EQ(JoinHash::calculate_radix_bits(0, 1, JoinMode::Inner), 0ul);
+  EXPECT_EQ(JoinHash::calculate_radix_bits(0, 0, JoinMode::Inner), 0ul);
+  EXPECT_EQ(JoinHash::calculate_radix_bits(1, 1, JoinMode::Inner), 0ul);
+  EXPECT_GT(JoinHash::calculate_radix_bits(std::numeric_limits<size_t>::max(), std::numeric_limits<size_t>::max(),
+                                           JoinMode::Inner),
+            0ul);
 }
 
 }  // namespace opossum
